fix failing unit tests
